Extract login and logout logic
