+++ client/src/App.tsx
@@ -80,6 +80,7 @@ function App(props: {}) {
       <main className="px-3 py-5">
         <Router>
           <Home path="/" />
+          {/* TODO: better page than homepage? Notification of logout success? */}
           <Home path="logout" />
           <LoginPage path="login" username={userInfo.username} setToken={setToken} />
           <SignupPage path="signup" username={userInfo.username} setToken={setToken} />
